# Enable the optimizing macros UNLIKELY, LIKELY, XML_ATTR_MALLOC and XML_ATTR_ALLOC_SIZE for clang-cl on Windows