--- a/Modules/expat/expat_external.h
+++ b/Modules/expat/expat_external.h
@@ -112,14 +112,14 @@
 #  define XMLIMPORT
 #endif
 
-#if defined(__GNUC__)                                                          \
+#if defined(__clang__) || defined(__GNUC__)                                    \
     && (__GNUC__ > 2 || (__GNUC__ == 2 && __GNUC_MINOR__ >= 96))
 #  define XML_ATTR_MALLOC __attribute__((__malloc__))
 #else
 #  define XML_ATTR_MALLOC
 #endif
 
-#if defined(__GNUC__)                                                          \
+#if defined(__clang__) || defined(__GNUC__)                                    \
     && ((__GNUC__ > 4) || (__GNUC__ == 4 && __GNUC_MINOR__ >= 3))
 #  define XML_ATTR_ALLOC_SIZE(x) __attribute__((__alloc_size__(x)))
 #else