--- a/Objects/obmalloc.c
+++ b/Objects/obmalloc.c
@@ -1457,7 +1457,7 @@ PyObject_Free(void *ptr)
 
 /* If we're using GCC, use __builtin_expect() to reduce overhead of
    the valgrind checks */
-#if defined(__GNUC__) && (__GNUC__ > 2) && defined(__OPTIMIZE__)
+#if (defined(__clang__) || (defined(__GNUC__) && (__GNUC__ > 2))) && defined(__OPTIMIZE__)
 #  define UNLIKELY(value) __builtin_expect((value), 0)
 #  define LIKELY(value) __builtin_expect((value), 1)
 #else