@chris-eibl 2025-03-09
Ups, [expat_external.h](https://github.com/python/cpython/pull/131019/files#diff-5ba2423c29b4fae208f4a55ee0875c6e2b2e8a50958786b533567a32d690a967) so this one should be left alone (or thouched upstream).